支払い機能の修正
